libproc: Add support for some proc_attach() flags.
This change adds some handling for the equivalent of Solaris' PGRAB_*
flags. In particular, support for PGRAB_RDONLY is needed to avoid a
nasty deadlock: dtrace(1) may otherwise stop the master process for its
pseudo-terminal and end up blocking while writing to standard output.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@309597 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ lib/libproc/proc_util.c
@@ -87,21 +87,25 @@ proc_detach(struct proc_handle *phdl, int reason)
 
 	if (phdl == NULL)
 		return (EINVAL);
+	if (reason == PRELEASE_HANG)
+		return (EINVAL);
 	if (reason == PRELEASE_KILL) {
 		kill(proc_getpid(phdl), SIGKILL);
-		return (0);
+		goto free;
 	}
+	if ((phdl->flags & PATTACH_RDONLY) != 0)
+		goto free;
 	pid = proc_getpid(phdl);
 	if (ptrace(PT_DETACH, pid, 0, 0) != 0 && errno == ESRCH)
-		return (0);
+		goto free;
 	if (errno == EBUSY) {
 		kill(pid, SIGSTOP);
 		waitpid(pid, &status, WUNTRACED);
 		ptrace(PT_DETACH, pid, 0, 0);
 		kill(pid, SIGCONT);
-		return (0);
 	}
-
+free:
+	proc_free(phdl);
 	return (0);
 }
